disp: msm: sde: Program RC data before enabling the DSPP RC feature
The DSPP's rounded-corner (RC) feature requires a large set of associated
data that must be programmed to the DSPP. To make this programming process
faster and more efficient, the hardware has an auto-incrementing index that
can be used with DMA to quickly copy all of the RC data at once, instead of
copying it one word at a time.

When post-processing features such as RC utilize DMA for programming
registers/RAMs, the actual programming over DMA is deferred until encoder
kickoff, at which point all of the DMA transfers are performed in a batch.
Meanwhile, the control registers for the DSPP's RC feature are programmed
immediately using plain MMIO writes, and they're programmed prior to the RC
data's programming, even when the RC data isn't programmed using DMA.

As it turns out, this creates a lethal race which is greatly exacerbated by
using DMA to transfer the RC data. If the DSPP's RC feature is enabled
without having valid RC data programmed, then ping-pong timeouts will
ensue, and the hardware won't recover without a soft reset:
  [  164.873972] [drm:_sde_encoder_phys_cmd_handle_ppdone_timeout:544] [sde error]enc55 intf1 pp:0 kickoff timed out ctl 0 koff_cnt 2
  [  164.914191] msm_drm ae00000.qcom,mdss_mdp: ctx:0, reg-dump logging time start_us:164822288, end_us:164849503 , duration_us:27214
  [  164.945471] msm_drm ae00000.qcom,mdss_mdp: pp0 0x348 33 0 0x8c7e
  [  164.952389] msm_drm ae00000.qcom,mdss_mdp: pp0 0x348 35 0 0x8000
  [  164.959378] msm_drm ae00000.qcom,mdss_mdp: pp0 0x348 42 0 0x8000
  [  164.966340] msm_drm ae00000.qcom,mdss_mdp: pp0 0x348 47 0 0x8000
  [  164.973317] msm_drm ae00000.qcom,mdss_mdp: pp0 0x348 49 0 0x8000
  [  164.980307] msm_drm ae00000.qcom,mdss_mdp: pp1 0x348 32 0 0xc12f
  [  164.987313] msm_drm ae00000.qcom,mdss_mdp: xbar 0x348 0 0 0x0
  [  164.994235] msm_drm ae00000.qcom,mdss_mdp: xin halt:0x3fff0000, pnd err:0x0, src err:0x0
  [  165.045784] msm_drm ae00000.qcom,mdss_mdp: debug-bus logging time start_us:164849514, end_us:164981099 , duration_us:131585
  [  165.045941] [drm:sde_encoder_phys_cmd_wait_for_tx_complete:1571] [sde error]failed wait_for_idle: -110

This is observed after powering on the display a few dozen times with
DSPP RC enabled and RC data programmed over DMA.

This issue is more prevalent when RC data is programmed using DMA because
of how the DMA transfer is deferred until encoder kickoff. This creates a
long window in which the RC feature is enabled (via the control registers)
without having valid RC data programmed.

The RC data must therefore be programmed *before* the RC feature is enabled
via the control registers. For the non-DMA case, this just involves moving
up the RC data programming so that it happens before the control registers
are configured.

For the DMA case, this is more complicated because of how DMA transfers are
deferred until encoder kickoff. The RC control registers must be programmed
in the same DMA payload as the RC data to ensure that they are configured
*after* RC data is programmed. The DMA case now does this by tacking on the
RC control registers' programming at the end of the RC data payload.

Change-Id: Ic53660320d86f0e2cff9a8684e52399c2b789545
Signed-off-by: Sultan Alsawaf <sultan@osomprivacy.com>